pyproject.toml
[tool.poetry]
name = "matbloom-api"
version = "0.1.0"
description = ""
authors = ["LazimR <lrodrigues356@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<4"
fastapi = ">=0.115.11,<0.116.0"
uvicorn = ">=0.34.0,<0.35.0"
sqlalchemy = ">=2.0.39,<3.0.0"
psycopg2-binary = ">=2.9.10,<3.0.0"
pydantic = {extras = ["email"], version = "^2.11.4"}
reportlab = ">=4.3.1,<5.0.0"
weasyprint = ">=65.0,<66.0"
pypdf2 = ">=3.0.1,<4.0.0"
numpy = ">=2.2.4,<3.0.0"
opencv-python = ">=4.11.0.86,<5.0.0.0"
pytesseract = ">=0.3.13,<0.4.0"
pytest = ">=8.3.5,<9.0.0"
httpx = ">=0.28.1,<0.29.0"
python-dotenv = ">=1.1.0,<2.0.0"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
passlib = "^1.7.4"
bcrypt = "^4.3.0"
python-multipart = "^0.0.20"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.requires-plugins]
poetry-plugin-export = "^1.7.1"